feat: add MID360S lidar driver support
