+++ livox_ros_driver2/.gitignore
@@ -1,3 +1,4 @@
 .vscode
 build
+package.xml
 __pycache__
