+++ livox_ros_driver2/README.md
@@ -13,6 +13,7 @@ Livox ROS Driver 2 is the 2nd-generation driver package used to connect LiDAR pr
   * Ubuntu 18.04 for ROS Melodic;
   * Ubuntu 20.04 for ROS Noetic and ROS2 Foxy;
   * Ubuntu 22.04 for ROS2 Humble;
+  * Ubuntu 24.04 for ROS 2 Jazzy;
 
   **Tips:**
 
@@ -33,6 +34,10 @@ For ROS2 Foxy installation, please refer to:
 
 For ROS2 Humble installation, please refer to:
 [ROS Humble installation instructions](https://docs.ros.org/en/humble/Installation/Ubuntu-Install-Debians.html)
+
+For ROS2 Jazzy installation, please refer to:
+[ROS Jazzy installation instructions](https://docs.ros.org/en/jazzy/Installation/Ubuntu-Install-Debians.html)
+
 
 Desktop-Full installation is recommend.
 
@@ -72,6 +77,13 @@ source /opt/ros/foxy/setup.sh
 ```shell
 source /opt/ros/humble/setup.sh
 ./build.sh humble
+```
+
+#### For ROS2 Jazzy:
+
+```shell
+source /opt/ros/jazzy/setup.sh
+./build.sh jazzy
 ```
 
 ### 2.4 Run Livox ROS Driver 2:
@@ -137,7 +149,7 @@ All internal parameters of Livox_ros_driver2 are in the launch file. Below are d
 
   **Note :**
 
-  Other parameters not mentioned in this table are not suggested to be changed unless fully understood.
+Other parameters not mentioned in this table are not suggested to be changed unless fully understood.
 
 &ensp;&ensp;&ensp;&ensp;***Livox_ros_driver2 pointcloud data detailed description :***
