+++ livox_ros_driver2/config/MID360_config.json
@@ -11,13 +11,13 @@
       "log_data_port": 56500
     },
     "host_net_info" : {
-      "cmd_data_ip" : "192.168.1.50",
+      "cmd_data_ip" : "192.168.1.5",
       "cmd_data_port": 56101,
-      "push_msg_ip": "192.168.1.50",
+      "push_msg_ip": "192.168.1.5",
       "push_msg_port": 56201,
-      "point_data_ip": "192.168.1.50",
+      "point_data_ip": "192.168.1.5",
       "point_data_port": 56301,
-      "imu_data_ip" : "192.168.1.50",
+      "imu_data_ip" : "192.168.1.5",
       "imu_data_port": 56401,
       "log_data_ip" : "",
       "log_data_port": 56501
@@ -25,7 +25,7 @@
   },
   "lidar_configs" : [
     {
-      "ip" : "192.168.1.3",
+      "ip" : "192.168.1.12",
       "pcl_data_type" : 1,
       "pattern_mode" : 0,
       "extrinsic_parameter" : {
