+++ livox_ros_driver2/src/call_back/livox_lidar_callback.cpp
@@ -24,6 +24,7 @@
 
 #include "livox_lidar_callback.h"
 
+#include "livox_lidar_api.h"
 #include <string>
 #include <thread>
 #include <iostream>
