+++ livox_ros_driver2/src/call_back/livox_lidar_callback.h
@@ -29,8 +29,6 @@
 #include "../lds_lidar.h"
 #include "../comm/comm.h"
 
-#include "livox_lidar_api.h"
-#include "livox_lidar_def.h"
 
 namespace livox_ros {
