+++ livox_ros_driver2/src/comm/pub_handler.cpp
@@ -23,7 +23,7 @@
 //
 
 #include "pub_handler.h"
-
+#include "livox_lidar_api.h"
 #include <cstdlib>
 #include <chrono>
 #include <iostream>
@@ -132,7 +132,7 @@ void PubHandler::OnLivoxLidarPointCloudCallback(uint32_t handle, const uint8_t d
   packet.extrinsic_enable = false; 
   if (dev_type == LivoxLidarDeviceType::kLivoxLidarTypeIndustrialHAP) {
     packet.line_num = kLineNumberHAP;
-  } else if (dev_type == LivoxLidarDeviceType::kLivoxLidarTypeMid360) {
+  } else if (dev_type == LivoxLidarDeviceType::kLivoxLidarTypeMid360||dev_type==LivoxLidarDeviceType::kLivoxLidarTypeMid360s) {
     packet.line_num = kLineNumberMid360;
   } else {
     packet.line_num = kLineNumberDefault;
