+++ livox_ros_driver2/src/comm/pub_handler.h
@@ -35,7 +35,6 @@
 #include <mutex>              // std::mutex
 #include <thread>
 
-#include "livox_lidar_def.h"
 #include "livox_lidar_api.h"
 #include "comm/comm.h"
