+++ livox_ros_driver2/src/include/livox_ros_driver2.h
@@ -27,7 +27,7 @@
 
 #define LIVOX_ROS_DRIVER2_VER_MAJOR 1
 #define LIVOX_ROS_DRIVER2_VER_MINOR 2
-#define LIVOX_ROS_DRIVER2_VER_PATCH 4
+#define LIVOX_ROS_DRIVER2_VER_PATCH 6
 
 #define GET_STRING(n) GET_STRING_DIRECT(n)
 #define GET_STRING_DIRECT(n) #n
